fixed a stupid random fill bug
Still haven't fount the main problem yet. Strangly enough, the inverse
overlapmatrix works, but the projection isn't orthogonal. (Just run the program and you'll see).

So probably what I see as the overlapmatrix isn't the overlapmatrix, or something else is going wrong.
Check tomorrow in Ward's program.

+++ LinIneq.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include <fstream>
+#include <cstdlib>
 #include <cmath>
 
 using std::cout;
@@ -591,8 +592,8 @@ void LinIneq::invert(){
 void LinIneq::fill_Random(){
 
    for(int i = 0;i < nr;++i)
-      proj[i] = rand()/RAND_MAX;
-
+      proj[i] = (double) rand()/RAND_MAX;
+   
 }
 
 /**
